fix: расширить TransactionId int -> long
ID транзакции (TRA_) на нагруженных серверах может превышать 2.1 млрд.
TransactionInfo.TransactionId и StyledProperty в 10 контролах EventCards
переведены на long; transaction_id парсится через ParseLongOrDefault.

+++ FirebirdTraceParser/Models/ValueObjects/TransactionInfo.cs
@@ -12,7 +12,7 @@ public sealed record TransactionInfo
 {
     [SortableField("Transaction ID", Category = "Transaction")]
     [FilterableField("Transaction ID", Category = "Transaction", FilterType =  FilterType.StringMultiSelect)]
-    public int? TransactionId { get; init; }
+    public long? TransactionId { get; init; }
     
     /// <summary>Уровень изоляции (READ_COMMITTED, SNAPSHOT, etc.)</summary>
     [SortableField("Isolation level", Category = "Transaction")]
